requirements.txt
# ─── Google ADK + Gemini ─────────────────────────────────────
google-adk>=1.31.1
google-generativeai>=0.8.6
google-genai>=1.73.1
litellm>=1.75.5,<=1.82.6

# ─── Google Cloud ────────────────────────────────────────────
google-cloud-firestore>=2.27.0
google-cloud-storage>=3.10.1
google-auth>=2.49.2
google-auth-oauthlib>=1.3.1
google-auth-httplib2>=0.3.1
google-api-python-client>=2.194.0

# ─── Web Framework ───────────────────────────────────────────
fastapi>=0.136.1
uvicorn[standard]>=0.46.0
python-multipart>=0.0.26

# ─── Data Validation ─────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.14.0

# ─── Research / Search ───────────────────────────────────────
duckduckgo-search>=8.1.1
httpx>=0.28.1
requests>=2.33.1
feedparser>=6.0.12
pytrends>=4.9.2

# ─── Optional Research (premium) ─────────────────────────────
tavily-python>=0.7.23

# ─── TTS / Audio ─────────────────────────────────────────────
edge-tts>=7.2.8

# ─── Captions (Whisper word-level subtitles) ─────────────────
openai-whisper>=20250625

# ─── Video / Image (production agent) ────────────────────────
Pillow>=12.2.0
ffmpeg-python>=0.2.0

# ─── Utils ───────────────────────────────────────────────────
python-dotenv>=1.0.1,<2.0.0
python-dateutil>=2.9.0.post0
tenacity>=9.1.4
structlog>=25.5.0
PyYAML>=6.0.3

# ─── Firebase Auth ───────────────────────────────────────────
firebase-admin>=7.4.0
